Move list_functions to update.py
